Move dependency groups support to `_internal`
- New CLI components are rehomed into `piptools._internal._cli`
- The `dependency-groups` module is moved to
  `piptools._internal._dependency_groups`
- Usage sites are updated to match
