+++ piptools/_internal/_cli/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from ._param_types import DependencyGroupParamType
+from ._parsed_param_types import ParsedDependencyGroupParam
+
+__all__ = (
+    "DependencyGroupParamType",
+    "ParsedDependencyGroupParam",
+)
